+++ data/src/main/java/io/tuttut/data/repository/diary/DiaryRepository.kt
@@ -1,12 +1,13 @@
 package io.tuttut.data.repository.diary
 
+import androidx.paging.PagingData
 import com.google.firebase.firestore.DocumentReference
 import io.tuttut.data.model.dto.Diary
 import kotlinx.coroutines.flow.Flow
 import io.tuttut.data.model.response.Result
 
 interface DiaryRepository {
-    fun getDiaryList(gardenId: String, cropsId: String): Flow<List<Diary>>
+    fun getDiaryList(gardenId: String, cropsId: String): Flow<PagingData<Diary>>
 
     fun getFourDiaryList(gardenId: String, cropsId: String): Flow<List<Diary>>
